--- src/pages/profile/Badges/BadgesSubojectives.module.css
@@ -1,28 +1,0 @@
-.BadgeProfileSubObjectivesFirst {
-    padding-top: 0.5rem;
-    padding-bottom: 0.5rem;
-    background-color: #292929;
-    margin-left: -0.9rem;
-    padding-left: 1rem;
-    margin-right: -0.9rem;
-}
-
-.BadgeProfileSubObjectivesMiddle {
-    padding-top: 0.5rem;
-    padding-bottom: 0.5rem;
-    background-color: #292929;
-    margin-left: -0.9rem;
-    padding-left: 1rem;
-    margin-right: -0.9rem;
-    margin-bottom: 1rem;
-    margin-top: 1rem;
-}
-
-.BadgeProfileSubObjectivesEnd {
-    padding-top: 0.5rem;
-    padding-bottom: 0.5rem;
-    background-color: #292929;
-    margin-left: -0.9rem;
-    padding-left: 1rem;
-    margin-right: -0.9rem;
-}
