--- src/pages/profile/Badges/NumberCompetencies.module.css
@@ -1,39 +1,0 @@
-.numberCompetenciesContent {
-    display: flex;
-    padding: 7px;
-    border: 2px solid rgb(163, 158, 158);
-    flex-grow: 1;
-    border-radius: 10px;
-    max-width: 50%;
-    position: static;
-    margin-left: 1rem;
-    padding-left: 15px;
-}
-
-.numberCompetenciesContentText {
-    display: flex;
-    flex-direction: column;
-    padding-left: 15px;
-}
-
-.numberCompetenciesContentText > span {
-    font-size: 1.3rem;
-    font-weight: 600;
-}
-
-
-/*
-apply a filter on the image
-.imageFilter {
-    filter: grayscale(80%);
-}*/
-
-.numberCompetenciesContentText > p {
-    color: #9c9c9c;
-}
-
-/*.numberCompetenciesContent > div {
-    display: flex;
-    padding : 3px;
-    flex-direction: column;
-}*/
