+++ src/pages/profile/statistics/badges/BadgesSubojectives.module.css
@@ -1,0 +1,5 @@
+.BadgeProfileSubObjectivesMiddle {
+    padding: 0.5rem 1rem;
+    background-color: #292929;
+    margin: 1rem -1rem 0;
+}
