+++ src/pages/profile/statistics/badges/index.tsx
@@ -1,0 +1,3 @@
+export {
+    BadgeSubojectives
+  } from "./BadgesSubojectives";
