+++ src/pages/profile/statistics/index.tsx
@@ -1,0 +1,1 @@
+export { Statistics } from "./Statistics"
