+++ src/pages/profile/statistics/progressBar/ProgressBar.module.css
