+++ src/pages/profile/statistics/progressBar/index.tsx
@@ -1,0 +1,1 @@
+export { ProgressBar } from "./ProgressBar"
